DataViews: Add `grid` keyboard navigation
